Add support for menu subsubitems (#1234)
This PR adds support for menu subsubitems (third level items).

## Screenshots


![image](https://github.com/vivid-planet/comet/assets/56400587/ee56d6b4-7417-44ed-9563-6a95fd8ba228)

![image](https://github.com/vivid-planet/comet/assets/56400587/d58686a0-b9ba-4eea-8185-bd24acc7dd0d)

---------
