+++ .changeset/pretty-deers-trade.md
@@ -1,0 +1,5 @@
+---
+"@comet/admin": minor
+---
+
+Add support for third level menu items. Collapsible items can be nested in each other, which creates subsubitems.
