+++ packages/admin/admin/src/index.ts
@@ -103,7 +103,8 @@ export { MasterLayoutClassKey } from "./mui/MasterLayout.styles";
 export { MasterLayoutContext } from "./mui/MasterLayoutContext";
 export { MenuCollapsibleItem, MenuCollapsibleItemClassKey, MenuCollapsibleItemProps, MenuLevel } from "./mui/menu/CollapsibleItem";
 export { IMenuContext, IWithMenu, MenuContext, withMenu } from "./mui/menu/Context";
-export { MenuItem, MenuItemClassKey, MenuItemProps } from "./mui/menu/Item";
+export { MenuItem, MenuItemProps } from "./mui/menu/Item";
+export { MenuItemClassKey } from "./mui/menu/Item.styles";
 export { MenuItemAnchorLink } from "./mui/menu/ItemAnchorLink";
 export { MenuItemGroup, MenuItemGroupClassKey, MenuItemGroupProps } from "./mui/menu/ItemGroup";
 export { MenuItemRouterLink, MenuItemRouterLinkProps } from "./mui/menu/ItemRouterLink";
