+++ packages/admin/admin/src/mui/menu/CollapsibleItem.tsx
@@ -28,7 +28,7 @@ const styles = (theme: Theme) =>
     });
 
 export interface MenuLevel {
-    level?: 1 | 2;
+    level?: 1 | 2 | 3;
 }
 
 type MenuChild = React.ReactElement<MenuItemRouterLinkProps>;
@@ -52,7 +52,7 @@ const CollapsibleItem: React.FC<WithStyles<typeof styles> & MenuCollapsibleItemP
     children,
     ...otherProps
 }) => {
-    const itemLevel: 1 | 2 = level ? level : 1;
+    const itemLevel: MenuLevel["level"] = level ? level : 1;
     let hasSelectedChild = false;
     const location = useLocation();
 
@@ -64,7 +64,7 @@ const CollapsibleItem: React.FC<WithStyles<typeof styles> & MenuCollapsibleItemP
         const newItemLevel = itemLevel + 1;
 
         return React.cloneElement<MenuLevel>(child, {
-            level: newItemLevel === 1 || newItemLevel === 2 ? newItemLevel : undefined,
+            level: newItemLevel === 1 || newItemLevel === 2 || newItemLevel === 3 ? newItemLevel : undefined,
         });
     });
 
@@ -80,6 +80,8 @@ const CollapsibleItem: React.FC<WithStyles<typeof styles> & MenuCollapsibleItemP
                 <MenuItem
                     primary={primary}
                     secondary={secondary}
+                    hasSubitems={!!childElements?.length}
+                    isCollapsibleOpen={open}
                     icon={icon}
                     level={level}
                     onClick={() => setOpen(!open)}
