+++ packages/api/cms-api/schema.gql
@@ -123,6 +123,11 @@ type FilenameResponse {
   isOccupied: Boolean!
 }
 
+type SitePreviewHash {
+  timestamp: Float!
+  hash: String!
+}
+
 type CurrentUserPermission {
   permission: String!
 }
@@ -153,11 +158,6 @@ enum UserPermissionSource {
 type PaginatedUserList {
   nodes: [User!]!
   totalCount: Int!
-}
-
-type SitePreviewHash {
-  timestamp: Float!
-  hash: String!
 }
 
 type PageTreeNode {
